Move bettiBoundary to utils, update correctly in fast persistence
